Move old code to a separate folder, fix on the fly download, move analysis
