+++ sponge/cli.py
@@ -1,55 +1,9 @@
 ### Imports ###
 from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
+
 from sponge import Sponge
-from sponge.config import DEFAULT_CHROMOSOMES, FILE_DF
-from typing import Dict, List
-from warnings import warn
 
 ### Functions ###
-def process_file_paths(
-    fp_list: List[str],
-) -> Dict[str, str]:
-    """
-    Converts a list of descriptor to path assignments to a dictionary,
-    invalid assignments result in a warning and are skipped.
-    Valid form: valid_descriptor=non-empty/path/to/file
-    Only one assignment operator (=) can be present.
-    Valid descriptors are: promoter, jaspar_bigbed, ensembl
-
-    Parameters
-    ----------
-    fp_list : List[str]
-        List of descriptor to path assignments to parse
-
-    Returns
-    -------
-    Dict[str, str]
-        Dictionary linking valid file descriptors to paths
-    """
-
-    processed = {}
-
-    for fp in fp_list:
-        split = fp.split('=')
-        if len(split) == 1:
-            warn('Path assignment does not contain an equal sign: '
-                f'{fp}, skipping')
-            continue
-        if len(split) > 2:
-            warn('Path assignment contains more than one equal sign: '
-                f'{fp}, skipping')
-            continue
-        if split[0] not in FILE_DF.index:
-            warn(f'File description not recognised: {split[0]}, skipping')
-            continue
-        if len(split[1]) == 0:
-            warn(f'Path to {split[0]} is empty, skipping')
-            continue
-        processed[split[0]] = split[1]
-
-    return processed
-
-
 def cli(
 ) -> None:
     """
@@ -63,11 +17,11 @@ def cli(
     Generates prior motif and PPI networks, usable by other NetZoo tools
     (most notably PANDA).
     Uses the Ensembl, JASPAR, NCBI and STRING databases.
-    Developed by Ladislav Hovan (ladislav.hovan@ncmm.uio.no).
+    Developed by Ladislav Hovan (ladislav.hovan@ncmbm.uio.no).
     """
     EPILOG = """
     Code available under GPL-3.0 license:
-    https://github.com/ladislav-hovan/sponge
+    https://github.com/kuijjerlab/sponge
     """
 
     parser = ArgumentParser(formatter_class=ArgumentDefaultsHelpFormatter,
@@ -98,7 +52,7 @@ def cli(
         action='store_true')
     parser.add_argument('-c', '--chromosomes', dest='chromosomes',
         help='chromosomes to select',
-        nargs='*', default=DEFAULT_CHROMOSOMES, metavar='CHR')
+        nargs='*', default=None, metavar='CHR')
     parser.add_argument('-to', '--tss-offset', dest='tss_offset',
         help='offset from TSS for promoters or regions of interest in which '
             'to search for TF binding sites',
@@ -132,25 +86,5 @@ def cli(
 
     args = parser.parse_args()
 
-    sponge_obj = Sponge(
-        temp_folder=args.temp_folder,
-        run_default=True,
-        jaspar_release=args.jaspar_release,
-        genome_assembly=args.genome_assembly,
-        n_processes=args.n_processes,
-        paths_to_files=process_file_paths(args.paths_to_files),
-        tf_names=args.tf_names,
-        matrix_ids=args.matrix_ids,
-        drop_heterodimers=(not args.keep_hd),
-        chromosomes=args.chromosomes,
-        tss_offset=tuple(args.tss_offset),
-        score_threshold=args.score_threshold,
-        on_the_fly_processing=args.on_the_fly,
-        protein_coding_only=args.pco,
-        use_gene_names=(not args.use_gene_ids),
-        weighted=args.weighted,
-        motif_outfile=args.motif_outfile,
-        ppi_outfile=args.ppi_outfile,
-        prompt=(not args.yes),
-    )
+    sponge_obj = Sponge()
     sponge_obj.show_fingerprint()
