+++ sponge/modules/analysis/__init__.py
@@ -1,0 +1,1 @@
+from sponge.modules.analysis.analysis import *
